+++ src/clj/rems/email/template.clj
@@ -137,17 +137,17 @@
                           :t.email.application-resubmitted/subject
                           :t.email.application-resubmitted/message)))
 
-(defmethod event-to-emails :application.event/comment-requested [event application]
-  (emails-to-recipients (:application/commenters event)
+(defmethod event-to-emails :application.event/review-requested [event application]
+  (emails-to-recipients (:application/reviewers event)
                         event application
-                        :t.email.comment-requested/subject
-                        :t.email.comment-requested/message))
+                        :t.email.review-requested/subject
+                        :t.email.review-requested/message))
 
-(defmethod event-to-emails :application.event/commented [event application]
+(defmethod event-to-emails :application.event/reviewed [event application]
   (emails-to-recipients (handlers application)
                         event application
-                        :t.email.commented/subject
-                        :t.email.commented/message))
+                        :t.email.reviewed/subject
+                        :t.email.reviewed/message))
 
 (defmethod event-to-emails :application.event/remarked [event application]
   (emails-to-recipients (handlers application)